Extracting handling of python "declarations" into `DeclarationHandler`
We consider some python nodes to be "declarations" and this PR splits off processing these into a dedicated `DeclarationHandler`, rather than having them in the `StatementHandler`.
